Add custom graphs on network size and relay stats.
We're now using Rserve to generate custom graphs on demand. Rserve runs in
the background and has all required R libraries and graph code already
loaded. When a client comes with a custom graph request, a Java servlet
passes the graph parameters to Rserve which writes a graph file to disk
that the servlet returns to the client. We're starting with graphs on
network size, relay versions, platforms, and observed bandwidth.

Also add logging and a property file.