Don't attempt to delete old graphs from the cache.
Tomcat shouldn't have permissions to delete old graph files from disk.
Instead, if a cached graph gets too old, we ask Rserve to overwrite it
with a new graph.

We could delete old graphs using a shell script, but it takes a long time
to fill a disk with graph files of about 25 KiB.