Stop serving JSPs directly and use servlets as controllers.
Note that these servlets are no real controllers yet (nor are the JSPs
just views). The new servlets simply forward requests to the JSPs that do
all the work. But in the future these servlets will request results from
model classes and ask JSPs to display them.

As a side-effect of adding servlets as controllers we don't have to
re-write URLs from *.jsp to *.html anymore.