Retire TemplateController and CustomGraphController.
The TemplateController class was used to assemble web pages consisting of
navigation bar, main body, and footer. We can simply use JSP's include
directive for this. Also move all JSPs to WEB-INF, so that they aren't
directly accessible for clients. Only the controller servlets are supposed
to access them when forwarding requests.

The CustomGraphController was used for the custom graphs page. Now that
all graphs can be customized and the custom graphs page has gone away, we
don't need the class anymore.