Added a Torstatus clone.
See the original Torstatus at torstatus.blutmagie.de. This version of torstatus
attempts to use the metrics-db database to portray a close-to-live view of the
Tor network from the consensus. It shows a sortable list of routers and various
router details, as well as a bandwidth graph for each router.