Improve relay search performance.
Our relay search consists of two nested SELECTs with the inner SELECT
returning the valid-after times of matching status entries. In that
statement, we shouldn't GROUP BY valid-after times and limit the result to
31 rows. We should rather return the last 31 valid-after times regardless
whether they're overlapping or not. This saves us a lot of join steps in
the outer SELECT resulting in rows that we're going to throw away anyway,
because the outer SELECT is limited to 31 rows, too.