Display all found relays of a consensus.
So far we limited search results to 30 matches. For a very broad search
(e.g. nicknames starting with 'U' like in 'Unnamed'), it's confusing that
we're only displaying the first 30 matches in a consensus. There's no way
to learn about the remaining matches in that consensus which is necessary
to narrow down the search.

Now we're displaying either all matches of a consensus or none (if we
already exceeded the 30 match limit).