Accept empty start or end dates.
So far we required that either both start and end date are provided, or
none of them.

Now we also accept the two cases when only one of the two dates is
provided.  If no end date is provided, we set it to today.  If no start
date is provided, we set it to 90 days before the end date.  This does
not affect the cases when both dates are provided (in which case we use
the provided dates), or when none are provided (in which case we use today
as the end date and today - 90 days as the start date).