Support Torperf's new .data format, too.
Torperf's new .data format has a 21st column that says whether the run
timeout out or not.  At the same time, the timestamp in columns 17 and 18
(datacompletesec and datacompleteusec) is set even in the timeout case.

We now support both .data formats (even in a single file) by checking how
many columns a row has and detecting timeouts accordingly.