Provide Torperf experiment tarballs.
Previously, we were listing the .data and .extradata files of the last
running Torperf experiment.  That's not really helpful for anyone.

Now, we provide all .data and .extradata files of a completed experiment
as a single tarball.

Only the default Torperf runs are still provided as single files, because
they are continuously updated.