Graph GetTor packages by language, not by bundle.
With the recent renaming of GetTor package names it has become more
difficult to decide which packages are a TBB and which are something else.
tor-browser-bundle_en is now one out of linux-i386_en, linux-x86_64_en,
macosx-i386_en, or windows_en.  And in the future there may be a TIMBB and
who knows what other packages.

Time to give up the guessing game.  The language code is something we can
identify for certain.  Let's graph that.