Rewrite the consensus-health checker.
The major change in this rewrite is that the consensus-health checker is
now independent from importing descriptors into the metrics database.  It
now downloads its own consensus and votes, triggered by its own cronjob.

This rewrite is the second in a series of refactorings to decouple the
many metrics-web components.  The first refactoring was to move ExoneraTor
data into its own database and use a separate database importer for it.
The goal will be to have separate component for each functionality without
dependencies, so that every component can fail independent from the rest.

Another effect of the rewrite is that the consensus-health checker can now
be extended more easily by adding new reports or new checks.