Stop archiving consensuses and votes after checking them.
The idea of archiving consensus and votes was that we could compare them
over time to see if, e.g., a consensus parameter gets added or removed.
This works in theory, but in practice, this approach makes the code more
complex than it needs to be.  We can as well change valid consensus
parameters in the code.