Redirect stderr, or it will be sent to the mailing list.
In theory, our cronjob shouldn't include stderr in whatever it pipes to
mail.  But Ant doesn't make a difference between stdout and stderr and
writes everything to stdout.  When we grep Ant's output we include the
stderr part, too.  Now we explicitly redirect stderr to a file.  We
usually don't need it, but when something goes wrong, we might want to
take a look.