Take out top-10 relays by bandwidth history.
Updating this table for a new time interval took 2--5 minutes.  That's
unacceptable.  This happened at least four times a day when the cached
table data got stale or whenever a user entered a different start or end
date of the time interval.  This doesn't just delay page load times, but
also keeps the database busy and affects other users.

If we want this table back, we'll have to pre-calculate some values.