Make R object generation thread-safe.
Problems with concurrent R object generation were highly unlikely before
this patch.  But that makes them even harder to track down.  Now, we're
generating all R objects in a separate thread that we join.  Before
starting a new thread we check if there's already a thread running to
generate the same object, and if so, we join that one.