Fix missing values in direct users graphs.
When we didn't have a single directory mirror reporting requests from a
given country, we treated these dates as "don't have any data" and
interrupted drawing the line.  This makes sense for times when there was a
general problem with statistics gathering in the network; we wouldn't want
all graphs to drop to zero in that case.  But if there was no such
network-wide event, we could interpret "no data for that country, but lots
of data for other countries" as "no users from that country" and continue
drawing the line at 0 users.

Fixes the rest of #6170.