Hide Papers page and prepare for removing it.
Now that tech reports have moved to research.tpo, we shouldn't keep a
second, outdated list of them on metrics.tpo.  Listing metrics-related
tech reports was the main purpose of that page.  The papers are already
linked from the Formats page, and the relevant blog posts have been
converted to Tor tech reports.

For now, the Papers page is only invisible, so that bookmarks still work.
Should be safe to take out the page entirely in a few months from now.

Implements #7327.