Provide all aggregate statistics on the website.
This commit adds an intermediate layer between aggregating data in
cronjobs and visualizing results in graphs and tables on the website.  All
data that is graphed or otherwise presented now come from 6 new .csv files
that are publicly available and not from the database that is only locally
available on the metrics server.  A major advantage of this change is that
people can easily plot their own graphs or even develop a prettier metrics
website without writing their own data aggregation code.