Remove non-updated relays-by-country graph.
In February 2013 we wrote that "updating by-country data has been
temporarily disabled as of February 2013, because it has become
prohibitively expensive.  This is not a bug, but a known problem that we
will work on as time permits."

So far time didn't permit.  Once it does we should put the graph back.
But for now, leaving a non-updated graph on the page makes us look bad.