Move Q-and-A about user statistics to a text file.
These questions and answers are likely read by less than 10% of visitors,
and the remaining 90% wonder what that wall of text is.  We should write a
more general Q-and-A section covering the entire website.  Whoever cares
about how user statistics are calculated can read the text file.

Tweak the Q's and A's a tiny bit while converting them to plain text.