Wrap form input elements in label elements.
Putting an input element inside a label element along with its text
label means that you can click on the text to activate the input
element. In some cases this does not change behavior much; e.g. you can
click on "Start date:" to focus the corresponding input field. But it
makes a big difference for checkboxes: you can click anywhere on the
text to toggle the checkbox.