Remove somewhat broken reset button.
Our reset button worked rather unintuitively.  When the user selects a
previously unselected option and hits update, the page will reload
with that option being selected.  Now, when the user hits reset,
nothing happens, because all options are still unchanged as compared
to what they were when reloading the page.  It seems we cannot fix
this without using JavaScript.  But it also seems that users don't
expect reset buttons these days, so let's just remove it rather than
confusing users with it.