Use metrics-lib's getHidserv*() in hidserv module.
When Metrics' hidserv module was written in March 2015, metrics-lib
did not support hidserv-* lines in extra-info descriptors, so it had
to implement its own parsing code and parse extra-info descriptors a
second time.

This has changed in December 2015 with metrics-lib 1.1.0 adding
"support for hidden-service statistics in extra-info descriptors".

It's time to replace that custom parsing code in Metrics' hidserv
module and use the methods provided by metrics-lib.

Implements #20510.