Revert "Add txtorcon to the development page."
This reverts commit 0a6320b68fe5ba0b1db6e49ac1c25aa343ad0210.

atagar points out, and meejah confirms, that txtorcon doesn't parse
most of the descriptors provided by CollecTor, except for
microdescriptors which it parses for its own purposes.  Sounds like
Python developers are better off using Stem for parsing descriptors,
and that's the main use case when they come to this page.