Skip long resource strings.
This patch fixes a bug where we'd consider two resource strings with
the same first 2048 characters as two distinct resource strings
internally, but which the database considers the same, because it only
stores the first 2048 characters.  In reality, these are just hacking
attempts or broken clients, so we can as well discard these lines
entirely and not bother any further.