Don't cut off any data from CSV files.
When graphing data from CSV files it's not our job to make sure the
data is stable enough to be graphed. That would mean that whoever uses
our CSV files directly would have to make that sure by themselves. If
data is too recent to be graphed, it shouldn't be included in the CSV
files. As a side effect this makes the graphing process a little
easier.