Use helper functions for date breaks and labels.
The main motivation behind this change is that we don't have to call a
helper function ourselves anymore on our data to determine useful
breaks and labels. Instead, we let ggplot2 call our functions to
determine breaks and labels for the displayed data.

Another motivation is that this removes a fair amount of code, and
that it's one step further towards piping everything using %>%.

Last but not least, it gives us even more control over date breaks and
labels.