Stop using y axis labels.
In most, if not all, cases it's obvious from the graph title what
we're plotting on the y axis. The main reason why we had a y axis name
was to include the unit. But we can as well add the unit to the
labels, similar to what we're already doing with percent. This saves
more room for the plot area, and it's easier to read.