Make userstats pgTAP test work again.
All calls of merge() exept for the first have been broken. The reason
is that we're creating a temporary table as part of that function,
which will only be dropped at the end of the session. But given that
all pgTAP tests run in a single session, that temporary table was not
dropped. The fix is to drop it explicitly after calling merge().

This change doesn't affect production, because we're only calling the
merge() function once per session in production.

It's not entirely clear why these tests worked in the beginning. My
current guess is that there was a PostgreSQL or pgTAP change regarding
temporary tables or running tests in a single session. And maybe that
change happened years ago and we just did not notice until now.

Fixes #24713.