Adds generate-news-json Ant task
This commit updates UpdateNews to use the existing News class and to
accept the output filename as an argument. It also enables pretty
printing the output to remove the additional manual step of piping the
output through jq. (See #23854 comment:9)

The new Ant task requires that the JAR be built prior to running. It
will update news.json in place.