Refactors news item presentation using JSTL
Instead of rendering HTML using News.formatAsTableRow(), News entries
and now handed directly to the JSP which uses JSTL to take care of
presentation.

The format of news.json is changed to make links into an array of
dictionaries containing targets and labels instead of HTML snippets and
an additional field, shortDescription, is added with a non-HTML short
description of the news item.