Simplify censorship detector.
As part of the censorship detector, we're currently fitting a normal
distribution to a set of approximately 50 values.

After talking to George Danezis, the original author of our censorship
detector, it's safe to approximate this step by computing the mean and
standard deviation of these values and using those to parameterize the
normal distribution.

A test run with today's numbers produced the exact same result.

The main benefit of this simplification is that it will be easier for
others to reproduce our numbers, even without a statistics library as
powerful as SciPy.