Remove inconsistency between bandwidth history graphs.
The number of bytes in all bandwidth histories differed from the sum
of those numbers by Exit/Guard flag. The difference came from relays
that reported bandwidth histories but that the directory authorities
did not list as running.

With this commit we stop producing bandwidth numbers that are not
broken down by Exit/Guard flag, and we sum up those columns for the
bandwidth graph.

Implements #26015.