Don't terminate the hidserv module when parsing descriptors.
This issue came up when reprocessing descriptor archives for #26022:
we'd terminate the hidserv module when running into an issue with
parsing descriptors, where a possible issue is a consensus without a
"bandwidth-weights" line.

This seems wrong. It's okay to print out a warning in this case. But
in general, we don't want to abort the entire module in this case.

Related to #26022.