Fix flaw in noise-removing code of hidserv module.
In the technical report that the hidserv module is based on we write:

"When processing hidden-service statistics, we need to handle the fact
that they have been obfuscated by relays. As first step, we're
attempting to remove the additive Laplace-distributed noise by
rounding up or down to the nearest multiple of bin_size. The idea is
that it's most likely that noise was added to the closest right side
of a bin than to the right side of another bin. In step two, we're
subtracting half of bin_size, because the relay added between 0 and
bin_size − 1 to the originally observed value."

However, our code has a flaw: we use integer truncation rather which
always rounds toward zero, whereas we really want to use the floor
function which rounds towards negative infinity.

The fix is to use Math.floorDiv() for the division rather than common
integer division and truncation.

Fixes #26022.