Compute percentiles in clients module using linear interpolation.
So far we computed percentiles in the censorship detector of the
clients module using our own formula that picked existing values as
median or quartiles.

Now we're using numpy.percentile which, by default, uses linear
interpolation when the desired quantile lies between two data points.

We're also removing unused imports of pylab and matplotlib in this
commit, which is not directly related.

Implements part of #26035.