Handle days without any successful measurements.
In the past, there have been days without a single OnionPerf
successful measurement to onion services. Despite the fact that we
seriously need better monitoring, we also need to treat these results
correctly. So far, we'd have computed median and quartile values of
0.0 seconds for these cases, which is just wrong. We need to treat
these days as missing values when it comes to medians or quartiles.
That's what we're fixing here.

Fixes #28136.