Remove Torperf/OnionPerf plots with all sources.
OnionPerf results look to be comparable over time, but between vantage
points there are systematic deltas between the results. The "all"
plots show rises and falls where they actually don't exist, it's just
that a particular vantage point was offline so the average of the two
remaining moves noticeably.

In this commit we remove the source parameter from these graphs and
always include all sources separately in the graph, but not a
combination of all measurements together.

Implements #28603.