Explain onion service traffic graphs a bit better.
As discussed on the ticket, we can explain better which cells are
counted for onion service traffic graphs: it's cells relayed by
rendezvous points.

It was also confusing to say "per day" in the graph description when
bandwidth is in gigabits per second. Removed that part.

Finally, there's now a glossary entry for "rendezvous points" that is
based on the glossary entry in torspec.

Fixes #26171.